feat: 상품 타이머 이벤트 핸들러 분리 및 통합
상품 관련 타이머 이벤트 핸들러를 `productTimerHandlers.ts` 파일로 분리하여 코드의 가독성과 유지보수성을 향상시켰습니다. `main.basic.ts`에서 타이머 관련 로직을 제거하고, `ProductSelector` 컴포넌트에서 타이머를 시작하도록 수정하였습니다. 이를 통해 타이머 기능이 보다 명확하게 관리될 수 있게 되었습니다.
